modules/resource/manager/stitching-entity/src/dummy/rms/requirements.txt
Flask==0.10.1
Flask-XML-RPC==0.1.2
Jinja2==2.7.2
M2Crypto==0.22.3
MarkupSafe==0.18
SQLAlchemy==0.9.3
Werkzeug==0.9.4
blinker==1.3
cffi==0.8.1
cryptography==0.2.2
flup==1.0.2
itsdangerous==0.23
lxml==3.3.3
pika==0.9.13
pyOpenSSL==0.14
pycparser==2.10
python-dateutil==2.2
pytz==2014.1.1
six==1.5.2
wsgiref==0.1.2
paste==1.7.5.1
